fixed circular reference
