+++ src/main/java/org/ohdsi/webapi/tag/TagService.java
@@ -35,11 +35,6 @@ public class TagService extends AbstractDaoService {
     private final TagRepository tagRepository;
     private final EntityManager entityManager;
     private final ConversionService conversionService;
-    private final PathwayService pathwayService;
-    private final CcService ccService;
-    private final CohortDefinitionService cohortDefinitionService;
-    private final ConceptSetService conceptSetService;
-    private final IRAnalysisService irAnalysisService;
 
     private final ArrayList<Producer<List<TagInfo>>> infoProducers;
 
@@ -47,20 +42,10 @@ public class TagService extends AbstractDaoService {
     public TagService(
             TagRepository tagRepository,
             EntityManager entityManager,
-            ConversionService conversionService,
-            PathwayService pathwayService,
-            CcService ccService,
-            CohortDefinitionService cohortDefinitionService,
-            ConceptSetService conceptSetService,
-            IRAnalysisService irAnalysisService) {
+            ConversionService conversionService) {
         this.tagRepository = tagRepository;
         this.entityManager = entityManager;
         this.conversionService = conversionService;
-        this.pathwayService = pathwayService;
-        this.ccService = ccService;
-        this.cohortDefinitionService = cohortDefinitionService;
-        this.conceptSetService = conceptSetService;
-        this.irAnalysisService = irAnalysisService;
 
         this.infoProducers = new ArrayList<>();
         this.infoProducers.add(tagRepository::findCohortTagInfo);
@@ -243,32 +228,6 @@ public class TagService extends AbstractDaoService {
             groupIds.add(g.getId());
             findParentGroup(g.getGroups(), groupIds);
         });
-    }
-
-    public void assignGroup(TagGroupSubscriptionDTO dto) {
-        if (isAdmin()) {
-            tagRepository.findByIdIn(new ArrayList<>(dto.getTags()))
-                    .forEach(tag -> {
-                        assignGroup(ccService, dto.getAssets().getCharacterizations(), tag.getId());
-                        assignGroup(pathwayService, dto.getAssets().getPathways(), tag.getId());
-                        assignGroup(cohortDefinitionService, dto.getAssets().getCohorts(), tag.getId());
-                        assignGroup(conceptSetService, dto.getAssets().getConceptSets(), tag.getId());
-                        assignGroup(irAnalysisService, dto.getAssets().getIncidenceRates(), tag.getId());
-                    });
-        }
-    }
-
-    public void unassignGroup(TagGroupSubscriptionDTO dto) {
-        if (isAdmin()) {
-            tagRepository.findByIdIn(new ArrayList<>(dto.getTags()))
-                    .forEach(tag -> {
-                        unassignGroup(ccService, dto.getAssets().getCharacterizations(), tag.getId());
-                        unassignGroup(pathwayService, dto.getAssets().getPathways(), tag.getId());
-                        unassignGroup(cohortDefinitionService, dto.getAssets().getCohorts(), tag.getId());
-                        unassignGroup(conceptSetService, dto.getAssets().getConceptSets(), tag.getId());
-                        unassignGroup(irAnalysisService, dto.getAssets().getIncidenceRates(), tag.getId());
-                    });
-        }
     }
 
     private <T extends Number> void assignGroup(HasTags<T> service, List<T> assetIds, Integer tagId) {
